spark-runtime: cross-check the declared SM count against the driver
`atlas_core::device::sm121::NUM_SMS = 48` -- a constant named after ONE part
-- was what every grid-sizing caller had, whatever hardware the build was for.
A build sized for the wrong number runs a grid that does not fit the device,
and nothing on either side ever said a number out loud, so it went unnoticed
for a whole campaign.

`kernels/<hw>/HARDWARE.toml` now declares `[hardware] sm_count` and build.rs
bakes it as `atlas_kernels::TARGET_SM_COUNT`. This adds the other half: the
boot preflight reads the driver's own
`CU_DEVICE_ATTRIBUTE_MULTIPROCESSOR_COUNT` and compares.

It WARNS and does not fail. A wrong count costs a suboptimal grid, never a
wrong answer, and refusing to boot on -- say -- an H100 PCIe with a different
bin would be a regression dressed as rigour. The warning names BOTH numbers
and the file the declaration came from, which is the line that would have
caught the defect in round 1. A driver that will not answer costs one
cross-check, not the boot: the arch verdict has already passed and the miss is
logged at debug.

`arch_preflight.rs` was at 493 of the 500-LoC cap, so its tests move to
`arch_preflight_tests.rs`. The split is by QUESTION -- the module states the
rules, the test file drives them against injected drivers -- so each file still
reads as one argument. The `DeviceQuery` seam gains `sm_count`, and the
existing call-sequence assertion now pins that the new query is addressed by
the SAME ordinal and runs after the arch verdict.
